chapter 2 peom
poem "সভ্যতার প্রতি" (Towards Civilization) by Rabindranath Tagore. In the poem, the speaker addresses "new civilization," calling it cruel and all-consuming. The speaker desires to return to a simpler, more natural way of life, asking for the forest, the shade of the hermitage, days free from sorrow, evening baths, cow grazing, peaceful Vedic hymns, humble food, and bark clothing. The speaker expresses a longing for inner contemplation and freedom, rejecting the "new luxuries in a stone cage" and instead wishing to regain personal strength and connect with the infinite heart of the world.